requirements.txt
flask==2.3.3
flask-socketio==5.3.6
eventlet==0.33.3
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
selenium==4.15.2
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
psutil==5.9.6
opencv-python==4.8.1.78
numpy==1.24.3
pyyaml==6.0.1
python-dotenv==1.0.0
ffmpeg-python==0.2.0
asyncio==3.4.3
aiohttp==3.9.0
websockets==12.0
